refactor , working sess and logout

+++ controllers/api/comment-routes.js
@@ -2,53 +2,49 @@ const router = require('express').Router();
 const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
-    Comment.findAll({
-        attributes: [
-            'id',
-            'post_id',
-            'user_id',
-            'comment_text'
-        ]
-    })
-    .then(dbPostData => res.json(dbPostData))
-      .catch(err => {
-        console.log(err);
-        res.status(500).json(err);
-      });
-  });
+  Comment.findAll()
+    .then(dbCommentData => res.json(dbCommentData))
+    .catch(err => {
+      console.log(err);
+      res.status(500).json(err);
+    });
+});
 
 router.post('/', (req, res) => {
+  // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
+  if (req.session) {
     Comment.create({
-        comment_text: req.body.comment_text,
-        user_id: req.body.user_id,
-        post_id: req.body.post_id
-      })
-        .then(dbCommentData => res.json(dbCommentData))
-        .catch(err => {
-          console.log(err);
-          res.status(400).json(err);
-        });
-
+      comment_text: req.body.comment_text,
+      user_id: req.session.user_id,
+      post_id: req.body.post_id
+    })
+      .then(dbCommentData => res.json(dbCommentData))
+      .catch(err => {
+        console.log(err);
+        res.status(400).json(err);
+      });
+  }
 });
 
 router.delete('/:id', (req, res) => {
+  if (req.session) {
     Comment.destroy({
-        where: {
-            id: req.params.id
-        }
+      where: {
+        id: req.params.id
+      }
     })
-    .then(dbPostData => {
-        if (!dbPostData) {
-          res.status(404).json({ message: 'No post found with this id' });
+      .then(dbCommentData => {
+        if (!dbCommentData) {
+          res.status(404).json({ message: 'No comment found with this id!' });
           return;
         }
-        res.json(dbPostData);
+        res.json(dbCommentData);
       })
       .catch(err => {
         console.log(err);
         res.status(500).json(err);
       });
-
+  }
 });
 
 module.exports = router;
